% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_symmetric.R
\name{el_symmetric}
\alias{el_symmetric}
\title{Elementary Symmetric Function
 }
\usage{
el_symmetric(values, k)
}
\description{
Evaluate elementary symmetric function, i.e., e(v,k) = e({x_1,x_2,x_3},1) = (x_1+x_2+x_3), or e({x_1,x_2,x_3},2) = (x_1*x_2+x_2*x_3+x_3*x_1)
 or,e({x_1,x_2,x_3},3) = (x_1*x_2*x_3).
}
\keyword{internal}
