% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{Gamma}
\alias{Gamma}
\title{Gamma}
\usage{
Gamma(Dy, alpha, prob.prior, weight.prior, mean.prior, sigma.prior,
  sigma.Dyo, prob.clutter, weights.clutter, mean.clutter, sigma.clutter,
  Nmax, b)
}
\arguments{
\item{x:}{two element vector. The posterior intensity is computed at x.}

\item{Dy:}{list of two element vectors representing points observed in a persistence diagram. Here we consider a fixed homological feature.
The coordinates (birth, death)/(birth, persistence) need to be a list of two element vectors.}

\item{alpha:}{The probability of a feature in the prior will be detected in the observation.}

\item{prob.prior:}{The prior cardinality is defined as a binomial and prob.prior is the probability term, i.e., prior cardinality ~ B(Nmax,prob.prior).}

\item{weight.prior:}{a list of mixture weights for the prior density object. This parameter will be an input for 'Wedge_Gaussian_Mixture' function to estimate prior density.}

\item{mean.prior:}{a list of two element vector, means of the prior density object. This parameter will be an input for 'Wedge_Gaussian_Mixture' function to estimate prior density.}

\item{sigma.prior:}{a list of positive constants, sigmas in covariance matrices of the prior density object. This parameter will be an input for 'Wedge_Gaussian_Mixture' function to estimate prior density.}

\item{sigma.Dyo:}{positive constant. variance coefficient of the likelihood density. This represents the degree of faith on the observed PDs representing the underlying truth.}

\item{prob.clutter:}{The clutter cardinality is defined as a binomial and prob.clutter is the probability term, i.e., clutter cardinality ~ B(Nmax,clutter).}

\item{weights.clutter:}{a list of mixture weights for the clutter density object. This parameter will be an input for 'Wedge_Gaussian_Mixture' function to estimate clutter density.}

\item{mean.clutter:}{a list of two element vector, means of the clutter density object. This parameter will be an input for 'Wedge_Gaussian_Mixture' function to estimate clutter density.}

\item{sigma.clutter:}{a list of positive constants, sigmas in covariance matrices of the clutter density object. This parameter will be an input for 'Wedge_Gaussian_Mixture' function to estimate clutter density.}

\item{Nmax:}{The maximum number of points at which the posterior cardinality will be truncated, i.e., {p_n} is computed for n=0 to n-Nmax. Also, this Nmax will be used to define prior and clutter cardinality. 
So, it should be large enough so that all Binomials involved in the computation make sense}

\item{b:}{0 or 1}
}
\value{
numeric
}
\description{
Gamma
}
\keyword{internal}
