% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaiid.R
\name{gammaiid}
\alias{gammaiid}
\title{gammaiid}
\usage{
gammaiid(Dy, alpha, prob.prior, weight.prior, mean.prior, sigma.prior,
  sigma.y, weights.unexpected, mean.unexpected, sigma.unexpected, Nmax, b)
}
\arguments{
\item{x:}{a 2 by 1 vector where the posterior intensity is computed. Cosequently, x is a coordinate in the wedge where the tilted PD is defined.}

\item{Dy:}{a list of n vectors(2 by 1) representing points observed in a tilted persistence diagram of a fixed homological feature.}

\item{alpha:}{0<=alpha<1. The probability of a feature in the prior will be detected in the observation.}

\item{prob.prior:}{The prior cardinality pmf is defined as a binomial and prob.prior is the probability term (\eqn{p_x} in Eqn.(2)).}

\item{weight.prior:}{a N by 1 vector of mixture weights (\eqn{c^X_{j}} of Eqn (1)) for the prior density estimation.}

\item{mean.prior:}{a list of N vectors(2 by 1) each represets mean of the prior density (\eqn{\mu^X_{j}} of Eqn (1))}

\item{sigma.prior:}{a N by 1 vector of positive constants, \eqn{\sigma^X_{j}} of Eqn (1).}

\item{sigma.y:}{a positive constant. Variance coefficient (\eqn{\sigma}) of the likelihood density \eqn{l(y|x)} defined in the description above. This represents the degree of faith on the observed PDs representing the prior.}

\item{weights.unexpected:}{a M by 1 vector of mixture weights for the unexpected features. i.e., (\eqn{c^Y_{i}} of Eqn (2) above.}

\item{mean.unexpected:}{a list of M vectors (2 by 1),each represets mean of the Gaussian mixture density (\eqn{\mu^Y_{i}} of Eqn (2)) for the unexpected features.}

\item{sigma.unexpected:}{a M by 1 vector of positive constants, \eqn{\sigma^Y_{i}} of Eqn (2).}

\item{Nmax:}{The maximum number of points on which the posterior cardinality will be truncated, i.e., {p_n} is computed for n=0 to n=Nmax. Also, this Nmax will be used to define prior cardinality too. 
So, it should be large enough so that all Binomials involved in the computation make sense}

\item{b:}{0 or 1}
}
\value{
numeric
}
\description{
gammaiid
}
\keyword{internal}
