% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_wedge_gaussian.R
\name{Wedge_Gaussian_evaluate}
\alias{Wedge_Gaussian_evaluate}
\title{Wedge Gaussian Mixture}
\arguments{
\item{x:}{two-element numeric}
}
\value{
numeric
}
\description{
Wedge Gaussian Mixture
}
