% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wedge_Gaussian_Mixture.R
\name{Wedge_Gaussian_Mixture}
\alias{Wedge_Gaussian_Mixture}
\title{Restricted Gaussian Mixure density}
\usage{
Wedge_Gaussian_Mixture(weights,mean,sigma)
}
\arguments{
\item{weights:}{a list of mixture weights}

\item{mean:}{a list of two element vector, means of the restricted Gaussian}

\item{sigma:}{a list of positive constants, sigmas in covariance matrices of the restricted Gaussian}
}
\value{
The function 'Wedge_Gaussian_Mixture' returns numerical value of the mixed Gaussian density restricted to wedge W.
}
\description{
Computes mixed Gaussian density restricted to wedge W. Require 'mvtnorm'package.
}
\examples{
# input the mean and constant sigma of covariance matrix
w <- Wedge_Gaussian_Mixture(weights = list(1), means = list(c(1,1)), sigmas= list(0.001))
w$evaluate(c(0.5,0.5))
}
