% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wedge_Gaussian_Mixture.R
\name{Wedge_Gaussian_Mixture}
\alias{Wedge_Gaussian_Mixture}
\title{Restricted Gaussian Mixture density}
\usage{
Wedge_Gaussian_Mixture(x, weights, means, sigmas)
}
\arguments{
\item{x:}{two element vector. These are the points to evaluate density at.}

\item{weights:}{a vector of the mixture weights}

\item{means:}{a list of two element vector where the vectors are the means of the restricted Gaussian}

\item{sigmas:}{a vector of positive constants, sigmas in covariance matrices of the restricted Gaussian}
}
\value{
The function \code{Wedge_Gaussian_Mixture} returns numerical value of the mixed Gaussian density restricted to wedge W.
}
\description{
Computes mixed Gaussian density restricted to wedge W. Require \code{mvtnorm} package.
}
\examples{
# input the mean and constant sigma of covariance matrix
w <- Wedge_Gaussian_Mixture(x = c(1,2), weights = 1, means = list(c(1,1)), sigmas= 0.1)
}
\keyword{internal}
